+++ src/entities/application/index.ts
@@ -1,3 +1,2 @@
-export { default as getMyApplication } from './api/getMyApplication';
 export * from './api/hooks';
 export * from './model/types';
