+++ src/entities/user/index.ts
@@ -1,3 +1,2 @@
-export { default as getMyInfo } from './api/getMyInfo';
 export * from './api/hooks';
 export * from './model/types';
