+++ src/widgets/applicationSection/index.ts
@@ -1,2 +1,1 @@
-export { default as ApplicationForm } from './ui/ApplicationForm';
 export { default as ApplicationSection } from './ui/ApplicationSection';
