+++ src/widgets/applyDepartment/index.ts
@@ -1,0 +1,1 @@
+export { default as ApplicationForm } from './ui/ApplicationForm';
